chore(proof): remove duplicate types from nitro-enclave in favor of base-proof-primitives
Delete types.rs entirely — Proposal, ProofJournal, TeeProofResult, and related
constants were duplicates of base-proof-primitives types. The enclave crate
now imports directly from base-proof-primitives instead of maintaining its
own copies. The print_real_config_hashes test moves to server.rs where the
related config hash tests already live.

Closes #1667

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-opencode)
