--- crates/proof/tee/nitro-enclave/src/types.rs
@@ -1,223 +1,0 @@
-//! Enclave-native proof types.
-//!
-//! Proof types mirror their counterparts in `base-proof-primitives` but are
-//! owned by the enclave crate so that changes to the host-side types (e.g.
-//! adding fields to `ProofRequest`) do not alter the enclave binary and
-//! therefore do not change the PCR0 measurement.
-//!
-//! Configuration types (`PerChainConfig`, `BlockId`, `Genesis`,
-//! `GenesisSystemConfig`) live in `base-proof-primitives` so they can be
-//! shared with zkVM programs without pulling in enclave-specific dependencies.
-
-use alloy_primitives::{Address, B256, Bytes};
-
-// ---------------------------------------------------------------------------
-// Proof types
-// ---------------------------------------------------------------------------
-
-/// ECDSA signature length in bytes (r: 32 + s: 32 + v: 1).
-pub const ECDSA_SIGNATURE_LENGTH: usize = 65;
-
-/// Base length of the proof journal without intermediate roots:
-/// address(20) + 5 × bytes32(32) + 2 × uint64(8) = 196 bytes.
-pub const PROOF_JOURNAL_BASE_LENGTH: usize = 196;
-
-/// The `AggregateVerifier` contract journal encoding.
-///
-/// Serializes proposal fields into the byte format expected by on-chain verification:
-///
-/// ```text
-/// prover(20) || l1OriginHash(32) || prevOutputRoot(32)
-///   || startingL2Block(8) || outputRoot(32) || endingL2Block(8)
-///   || intermediateRoots(32*N) || configHash(32) || imageHash(32)
-/// ```
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ProofJournal {
-    /// The proposer address.
-    pub proposer: Address,
-    /// The L1 origin block hash.
-    pub l1_origin_hash: B256,
-    /// The previous output root hash.
-    pub prev_output_root: B256,
-    /// The starting L2 block number.
-    pub starting_l2_block: u64,
-    /// The output root hash.
-    pub output_root: B256,
-    /// The ending L2 block number.
-    pub ending_l2_block: u64,
-    /// Intermediate output roots for aggregate proposals.
-    pub intermediate_roots: Vec<B256>,
-    /// The config hash.
-    pub config_hash: B256,
-    /// The TEE image hash.
-    pub tee_image_hash: B256,
-}
-
-impl ProofJournal {
-    /// Encode the journal into the ABI-packed byte format.
-    #[must_use]
-    pub fn encode(&self) -> Vec<u8> {
-        let mut data =
-            Vec::with_capacity(PROOF_JOURNAL_BASE_LENGTH + 32 * self.intermediate_roots.len());
-
-        data.extend_from_slice(self.proposer.as_slice());
-        data.extend_from_slice(self.l1_origin_hash.as_slice());
-        data.extend_from_slice(self.prev_output_root.as_slice());
-        data.extend_from_slice(&self.starting_l2_block.to_be_bytes());
-        data.extend_from_slice(self.output_root.as_slice());
-        data.extend_from_slice(&self.ending_l2_block.to_be_bytes());
-        for root in &self.intermediate_roots {
-            data.extend_from_slice(root.as_slice());
-        }
-        data.extend_from_slice(self.config_hash.as_slice());
-        data.extend_from_slice(self.tee_image_hash.as_slice());
-
-        data
-    }
-}
-
-/// A proposal containing an output root and signature.
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-pub struct Proposal {
-    /// The output root hash.
-    pub output_root: B256,
-    /// The ECDSA signature (65 bytes: r, s, v).
-    pub signature: Bytes,
-    /// The L1 origin block hash.
-    pub l1_origin_hash: B256,
-    /// The L1 origin block number.
-    pub l1_origin_number: u64,
-    /// The L2 block number (ending block of this proposal's range).
-    pub l2_block_number: u64,
-    /// The previous output root hash.
-    pub prev_output_root: B256,
-    /// The config hash.
-    pub config_hash: B256,
-}
-
-/// Result of a TEE proof computation.
-///
-/// This is the enclave-native equivalent of `ProofResult::Tee` from
-/// `base-proof-primitives`, but as a flat struct rather than an enum variant.
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-pub struct TeeProofResult {
-    /// The aggregated proposal covering the entire proven block range.
-    pub aggregate_proposal: Proposal,
-    /// The individual per-block proposals that were aggregated.
-    pub proposals: Vec<Proposal>,
-}
-
-#[cfg(test)]
-mod tests {
-    use alloy_primitives::{address, b256};
-    use base_alloy_chains::BaseChainConfig;
-    use base_consensus_registry::Registry;
-    use base_proof_primitives::PerChainConfig;
-
-    use super::*;
-
-    fn test_journal() -> ProofJournal {
-        ProofJournal {
-            proposer: address!("f39Fd6e51aad88F6F4ce6aB8827279cffFb92266"),
-            l1_origin_hash: b256!(
-                "2222222222222222222222222222222222222222222222222222222222222222"
-            ),
-            prev_output_root: b256!(
-                "3333333333333333333333333333333333333333333333333333333333333333"
-            ),
-            starting_l2_block: 999,
-            output_root: b256!("4444444444444444444444444444444444444444444444444444444444444444"),
-            ending_l2_block: 1000,
-            intermediate_roots: vec![],
-            config_hash: b256!("1111111111111111111111111111111111111111111111111111111111111111"),
-            tee_image_hash: b256!(
-                "5555555555555555555555555555555555555555555555555555555555555555"
-            ),
-        }
-    }
-
-    #[test]
-    fn test_journal_encode_length() {
-        let data = test_journal().encode();
-        assert_eq!(data.len(), PROOF_JOURNAL_BASE_LENGTH);
-        assert_eq!(data.len(), 196);
-    }
-
-    #[test]
-    fn test_journal_encode_components() {
-        let journal = test_journal();
-        let data = journal.encode();
-
-        let mut off = 0;
-        assert_eq!(
-            &data[off..off + 20],
-            address!("f39Fd6e51aad88F6F4ce6aB8827279cffFb92266").as_slice()
-        );
-        off += 20;
-        assert_eq!(&data[off..off + 32], journal.l1_origin_hash.as_slice());
-        off += 32;
-        assert_eq!(&data[off..off + 32], journal.prev_output_root.as_slice());
-        off += 32;
-        assert_eq!(&data[off..off + 8], &journal.starting_l2_block.to_be_bytes());
-        off += 8;
-        assert_eq!(&data[off..off + 32], journal.output_root.as_slice());
-        off += 32;
-        assert_eq!(&data[off..off + 8], &journal.ending_l2_block.to_be_bytes());
-        off += 8;
-        assert_eq!(&data[off..off + 32], journal.config_hash.as_slice());
-        off += 32;
-        assert_eq!(&data[off..off + 32], journal.tee_image_hash.as_slice());
-    }
-
-    #[test]
-    fn test_journal_encode_with_intermediate_roots() {
-        let journal = ProofJournal {
-            proposer: address!("f39Fd6e51aad88F6F4ce6aB8827279cffFb92266"),
-            l1_origin_hash: B256::ZERO,
-            prev_output_root: B256::ZERO,
-            starting_l2_block: 0,
-            output_root: B256::ZERO,
-            ending_l2_block: 0,
-            intermediate_roots: vec![
-                b256!("aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"),
-                b256!("bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"),
-            ],
-            config_hash: B256::ZERO,
-            tee_image_hash: B256::ZERO,
-        };
-
-        let data = journal.encode();
-        assert_eq!(data.len(), PROOF_JOURNAL_BASE_LENGTH + 64);
-
-        let ir_offset = 20 + 32 + 32 + 8 + 32 + 8;
-        assert_eq!(&data[ir_offset..ir_offset + 32], journal.intermediate_roots[0].as_slice());
-        assert_eq!(&data[ir_offset + 32..ir_offset + 64], journal.intermediate_roots[1].as_slice());
-    }
-
-    /// Print config hashes for supported chains so they can be hardcoded in the
-    /// enclave server. Run with:
-    /// `cargo test -p base-proof-tee-nitro-enclave print_real_config_hashes -- --nocapture --ignored`
-    #[test]
-    #[ignore]
-    fn print_real_config_hashes() {
-        for cfg in BaseChainConfig::all() {
-            let chain_id = cfg.chain_id;
-            let rollup = match Registry::rollup_config(chain_id) {
-                Some(r) => r,
-                None => {
-                    println!("chain {chain_id}: skipped (no rollup config)");
-                    continue;
-                }
-            };
-            let mut per_chain = match PerChainConfig::from_rollup_config(rollup) {
-                Some(pc) => pc,
-                None => {
-                    println!("chain {chain_id}: skipped (no system_config)");
-                    continue;
-                }
-            };
-            per_chain.force_defaults();
-            println!("chain {chain_id}: {:?}", per_chain.hash());
-        }
-    }
-}
